Move TableBody to separate component
